Big changes in application, added error screen and more
