+++ quinzena8/astromatch/astromatch/src/components/Global/Toast/index.js
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import {ProfileContext} from '../../contexts/ProfileContext'
+import {ProfileContext} from '../../../contexts/ProfileContext'
 import { Container, Wrapper, Message } from './style'
 
 export function Toast() {
