+++ quinzena8/astromatch/astromatch/src/components/HomeComponents/NoProfile/index.js
@@ -1,0 +1,11 @@
+import Caring from '../../assets/caring.png'
+import { AlertWrapper, AlertImage, AlertMessage } from './style'
+
+export function NoProfile() {
+    return (
+        <AlertWrapper>
+            <AlertImage src={Caring} alt='Image'/>
+            <AlertMessage>Oops, os perfis <strong>acabaram</strong>! Por favor, clique no botão de reset.</AlertMessage>
+        </AlertWrapper>
+    )
+}
