+++ quinzena8/astromatch/astromatch/src/pages/HomePage/style.js
@@ -1,5 +1,4 @@
-import styled, { keyframes, css } from 'styled-components'
-import { generalColors } from '../../constants/colors'
+import styled from 'styled-components'
 import { Link } from "react-router-dom"
 
  
@@ -69,115 +68,6 @@ export const CardWrapper = styled.div`
     justify-content: center;
     align-items: center;
 `
-export const ProfileCard = styled.div`
-    width: 95%;
-    height: 100%;
-    background-image: ${props => `url(${props.profilePhoto})`};
-    background-position: center top;
-    background-size: cover;
-    display: flex;
-    align-items: flex-end;
-    border-radius: 8px;
-`
-export const ProfileInfo = styled.div`
-    width: 100%;
-    height: 65%;
-    transition: background 300ms linar;
-    background: ${p => p.theme.id === 1 ? `linear-gradient(transparent 65%, ${p.theme.backgroundWrapper} 78%)` : `linear-gradient(transparent 30%, ${p.theme.backgroundWrapper} 80%)`};
-    display: flex;
-    flex-flow: column;
-    justify-content: flex-end;
-    align-items: flex-start;
-    color: ${p => p.theme.text};
-    padding: 0 10px;
-`
-export const Name = styled.span`
-    font-weight: 600;
-    font-size: 18.5px;
-`
-export const Age = styled.span`
-    margin-left: 8px;
-    font-weight: 500;
-    font-size: 14px;
-`
-export const Activity = styled.div`
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    background-color: ${props => props.profileAge % 2 === 0 ? `${generalColors.green}` : `${generalColors.red}`};
-    display: inline-block;
-    margin-left: 5px;
-`
-export const Log = styled.span`
-    margin-left: 3px;
-    font-weight: 500;
-    font-size: 10px;
-`
-export const ProfileDescription = styled.p`
-    font-size: 12px;
-    color: ${p => p.theme.secondaryText};
-
-`
-export const AlertMessage = styled.p`
-    text-align: center;
-    font-size: 14px;
-    width: 90%;
-    color: ${p => p.theme.text};
-`
-export const AlertWrapper = styled.div`
-    display: flex;
-    flex-flow: column nowrap;
-    align-items: center;
-    justify-content: center;
-`
 export const AlertImage = styled.img`
     max-width: 180px;
-`
-
-// Footer
-export const Footer = styled.div`
-    grid-area: Footer;
-    border-radius: 0 0 10px 10px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-`
-export const Reload = styled.button`
-    border: none;
-    width: 50px;
-    height: 50px;
-    background-color: ${p => p.theme.button};
-    border-radius: 50%;
-    display: grid;
-    place-items: center;
-    font-size: 30px;
-    color: ${p => p.theme.text};
-    cursor: pointer;
-    &:hover {
-        filter: brightness(.95);
-    }
-    &:active {
-        transform: scale(0.95);
-    }
-`
-export const Dislike = styled(Reload)`
-    width: 50px;
-    height: 50px;
-`
-const pulseTap = keyframes`
-    0% {
-        border: 10px solid #FE1A1A;
-    }
-
-    100% {
-        border: 0px solid #FE1A1A00;
-    }
-`
-export const Like = styled(Dislike)`
-    width: 70px;
-    height: 70px;
-    color: ${generalColors.red};
-    margin: 0 10px;
-    font-size: 40px;
-    animation: ${p => p.handleLike ? css`${pulseTap} 500ms` : null};
 `
